float para int

+++ Lista de Charles/q1.py
@@ -5,7 +5,7 @@ Massa Inicial: 250 gramas
 Massa Final: 0.48828125 gramas
 Tempo de Decaimento: 0:07:30'''
 
-massaInicial = float(input("Digie a massa inicial do material radioativo: "))
+massaInicial = int(input("Digie a massa inicial do material radioativo: "))
 massa = massaInicial
 tempoEmSegundos = 0
